chore: update lib.rs file and adjust use declarations
